Renames ./file/readwrite/ --> ./file/io/
