+++ py4DSTEM/file/__init__.py
@@ -1,3 +1,3 @@
 from .log import *
 from . import datastructure
-from . import readwrite
+from . import io
