+++ py4DSTEM/file/io/append.py
